ref(derived): Move framework-only tests into their own file
